+++ prompt2model/dataset_processer/__init__.py
@@ -1,5 +1,5 @@
 """Import DatasetRetriever classes."""
-from prompt2model.dataset_processer.base import BasePrcesser
-from prompt2model.dataset_processer.mock import MockPrcesser
+from prompt2model.dataset_processor.base import BasePrcesser
+from prompt2model.dataset_processor.mock import MockProcessor
 
-__all__ = ("BasePrcesser", "MockPrcesser")
+__all__ = ("BasePrcesser", "MockProcessor")
